fixed exercise selector

+++ config.php
@@ -24,7 +24,7 @@ $db->exec("CREATE TABLE IF NOT EXISTS sets (
     user_id INT(11) NOT NULL,
     exercise_id INT(11) NOT NULL,
     reps INT(11) NOT NULL,
-    weight INT(11) NOT NULL,
+    weight DECIMAL(10, 3) NOT NULL,
     date DATE NOT NULL,
     PRIMARY KEY (id),
     FOREIGN KEY (user_id) REFERENCES users(user_id),
